Moved legacy tests to tests folder
